sp-de-technical-test/requirements.txt
SQLAlchemy~=1.3
psycopg2-binary~=2.8
sqlbucket==0.3.2
requests~=2.24
pandas == 1.5.3